+++ releases.md
@@ -25,6 +25,14 @@ Changes are grouped as follows:
 - OOTB incremental read support for time series.
 - Geo-location attribute and resource type.
 
+## [0.9.30-SNAPSHOT]
+
+### Added
+
+- `Read direct` mode for `Raw`. This read mode will output batches of rows (`Iterable<RawRow>`) instead of single 
+rows. This can be useful in very high data volume scenarios. 
+
+
 ## [0.9.29] 2021-11-09
 
 ### Changed